Allows deletion of ProjectUser memberships when they have a bad apikey.

+++ applications/wallet/app/projectusers/views.py
@@ -171,24 +171,23 @@ class ProjectUserViewSet(BaseProjectViewSet):
                             status=status.HTTP_403_FORBIDDEN)
 
         # Delete Users Apikey
+        apikey_readable = True
         try:
             key_data = decode_apikey(apikey)
-        except ValueError:
-            return Response(data={'detail': 'Unable to parse apikey.'},
-                            status=status.HTTP_400_BAD_REQUEST)
-        try:
             apikey_id = key_data['id']
-        except KeyError:
-            return Response(data={'detail': 'Apikey is incomplete.'},
-                            status=status.HTTP_400_BAD_REQUEST)
-        try:
-            response = client.delete(f'/auth/v1/apikey/{apikey_id}')
-        except ZmlpInvalidRequestException:
-            return Response(data={'detail': "Unable to delete apikey."},
-                            status=status.HTTP_400_BAD_REQUEST)
-        if not response.status_code == 200:
-            return Response(data={'detail': 'Error deleting apikey.'},
-                            status=status.HTTP_500_INTERNAL_SERVER_ERROR)
+        except (ValueError, KeyError):
+            print(f'Unable to decode apikey during delete for user {membership.user.id}.')
+            apikey_readable = False
+
+        if apikey_readable:
+            try:
+                response = client.delete(f'/auth/v1/apikey/{apikey_id}')
+            except ZmlpInvalidRequestException:
+                return Response(data={'detail': "Unable to delete apikey."},
+                                status=status.HTTP_400_BAD_REQUEST)
+            if not response.status_code == 200:
+                return Response(data={'detail': 'Error deleting apikey.'},
+                                status=status.HTTP_500_INTERNAL_SERVER_ERROR)
 
         membership.delete()
         return Response(status=status.HTTP_200_OK)
